+++ src/Storage/Storage.Management/StorageAccount/NewAzureStorageAccount.cs
@@ -38,6 +38,11 @@ namespace Microsoft.Azure.Commands.Management.Storage
         /// </summary>
         private const string ActiveDirectoryDomainServicesForFileParameterSet = "ActiveDirectoryDomainServicesForFile";
 
+        /// <summary>
+        /// Set AzureActiveDirectoryKerberosForFile parameter set name
+        /// </summary>
+        private const string AzureActiveDirectoryKerberosForFileParameterSet = "AzureActiveDirectoryKerberosForFile";        
+
         [Parameter(
             Position = 0,
             Mandatory = true,
@@ -203,6 +208,40 @@ namespace Microsoft.Azure.Commands.Management.Storage
 
         [Parameter(
             Mandatory = false,
+            HelpMessage = "Enable Secure File Transfer Protocol for the Storage account.")]
+        [ValidateNotNullOrEmpty]
+        public bool EnableSftp
+        {
+            get
+            {
+                return enableSftp != null ? enableSftp.Value : false;
+            }
+            set
+            {
+                enableSftp = value;
+            }
+        }
+        private bool? enableSftp = null;
+
+        [Parameter(
+            Mandatory = false,
+            HelpMessage = "Enable local users feature for the Storage account.")]
+        [ValidateNotNullOrEmpty]
+        public bool EnableLocalUser
+        {
+            get
+            {
+                return enableLocalUser != null ? enableLocalUser.Value : false;
+            }
+            set
+            {
+                enableLocalUser = value;
+            }
+        }
+        private bool? enableLocalUser = null;
+
+        [Parameter(
+            Mandatory = false,
             HelpMessage = "Enable HierarchicalNamespace for the Storage account.")]
         [ValidateNotNullOrEmpty]
         public bool EnableHierarchicalNamespace
@@ -280,6 +319,23 @@ namespace Microsoft.Azure.Commands.Management.Storage
         }
         private bool? publishInternetEndpoint = null;
 
+        [Parameter(
+            Mandatory = true,
+            HelpMessage = "Enable Azure Files Active Directory Domain Service Kerberos Authentication for the storage account.",
+            ParameterSetName = AzureActiveDirectoryKerberosForFileParameterSet)]
+        [ValidateNotNullOrEmpty]
+        public bool EnableAzureActiveDirectoryKerberosForFile
+        {
+            get
+            {
+                return enableAzureActiveDirectoryKerberosForFile.HasValue ? enableAzureActiveDirectoryKerberosForFile.Value : false;
+            }
+            set
+            {
+                enableAzureActiveDirectoryKerberosForFile = value;
+            }
+        }
+        private bool? enableAzureActiveDirectoryKerberosForFile = null;
 
         [Parameter(
             Mandatory = false,
@@ -301,8 +357,12 @@ namespace Microsoft.Azure.Commands.Management.Storage
 
         [Parameter(
             Mandatory = false,
-            HelpMessage = "Specifies the primary domain that the AD DNS server is authoritative for. This parameter must be set when -EnableActiveDirectoryDomainServicesForFile is set to true.",
+            HelpMessage = "Specifies the primary domain that the AD DNS server is authoritative for. This parameter must be set when -EnableActiveDirectoryDomainServicesForFile or -EnableAzureActiveDirectoryKerberosForFile is set to true.",
             ParameterSetName = ActiveDirectoryDomainServicesForFileParameterSet)]
+        [Parameter(
+            Mandatory = false,
+            HelpMessage = "Specifies the primary domain that the AD DNS server is authoritative for. This parameter must be set when -EnableActiveDirectoryDomainServicesForFile or -EnableAzureActiveDirectoryKerberosForFile is set to true.",
+            ParameterSetName = AzureActiveDirectoryKerberosForFileParameterSet)]
         [ValidateNotNullOrEmpty]
         public string ActiveDirectoryDomainName { get; set; }
 
@@ -322,8 +382,12 @@ namespace Microsoft.Azure.Commands.Management.Storage
 
         [Parameter(
             Mandatory = false,
-            HelpMessage = "Specifies the domain GUID. This parameter must be set when -EnableActiveDirectoryDomainServicesForFile is set to true.",
+            HelpMessage = "Specifies the domain GUID. This parameter must be set when -EnableActiveDirectoryDomainServicesForFile or -EnableAzureActiveDirectoryKerberosForFile is set to true.",
             ParameterSetName = ActiveDirectoryDomainServicesForFileParameterSet)]
+        [Parameter(
+            Mandatory = false,
+            HelpMessage = "Specifies the domain GUID. This parameter must be set when -EnableActiveDirectoryDomainServicesForFile or -EnableAzureActiveDirectoryKerberosForFile is set to true.",
+            ParameterSetName = AzureActiveDirectoryKerberosForFileParameterSet)]
         [ValidateNotNullOrEmpty]
         public string ActiveDirectoryDomainGuid { get; set; }
 
@@ -539,6 +603,11 @@ namespace Microsoft.Azure.Commands.Management.Storage
         [ValidateNotNullOrEmpty]
         public string ImmutabilityPolicyState { get; set; }
 
+        [Parameter(Mandatory = false, HelpMessage = "Set restrict copy to and from Storage Accounts within an AAD tenant or with Private Links to the same VNet. Possible values include: 'PrivateLink', 'AAD'")]
+        [PSArgumentCompleter("PrivateLink", "AAD")]
+        [ValidateNotNullOrEmpty]
+        public string AllowedCopyScope { get; set; }
+
         public override void ExecuteCmdlet()
         {
             base.ExecuteCmdlet();
@@ -608,7 +677,7 @@ namespace Microsoft.Azure.Commands.Management.Storage
             {
                 createParameters.IsHnsEnabled = enableHierarchicalNamespace;
             }
-            if (enableAzureActiveDirectoryDomainServicesForFile !=null || enableActiveDirectoryDomainServicesForFile != null)
+            if (enableAzureActiveDirectoryDomainServicesForFile !=null || enableActiveDirectoryDomainServicesForFile != null || enableAzureActiveDirectoryKerberosForFile != null)
             {
                 createParameters.AzureFilesIdentityBasedAuthentication = new AzureFilesIdentityBasedAuthentication();
                 if (enableAzureActiveDirectoryDomainServicesForFile != null && enableAzureActiveDirectoryDomainServicesForFile.Value)
@@ -640,6 +709,18 @@ namespace Microsoft.Azure.Commands.Management.Storage
                         SamAccountName = this.ActiveDirectorySamAccountName,
                         AccountType = this.ActiveDirectoryAccountType
                     };
+                }
+                else if (enableAzureActiveDirectoryKerberosForFile != null && enableAzureActiveDirectoryKerberosForFile.Value)
+                {
+                    createParameters.AzureFilesIdentityBasedAuthentication.DirectoryServiceOptions = DirectoryServiceOptions.AADKERB;
+                    if (this.ActiveDirectoryDomainName != null || this.ActiveDirectoryDomainGuid != null)
+                    {
+                        createParameters.AzureFilesIdentityBasedAuthentication.ActiveDirectoryProperties = new ActiveDirectoryProperties()
+                        {
+                            DomainName = this.ActiveDirectoryDomainName,
+                            DomainGuid = this.ActiveDirectoryDomainGuid
+                        };
+                    }
                 }
                 else
                 {
@@ -788,6 +869,18 @@ namespace Microsoft.Azure.Commands.Management.Storage
                     createParameters.ImmutableStorageWithVersioning.ImmutabilityPolicy.ImmutabilityPeriodSinceCreationInDays = this.immutabilityPeriod;
                     createParameters.ImmutableStorageWithVersioning.ImmutabilityPolicy.State = this.ImmutabilityPolicyState;
                 }
+            }
+            if (this.enableSftp != null)
+            {
+                createParameters.IsSftpEnabled = this.enableSftp;
+            }
+            if (this.enableLocalUser != null)
+            {
+                createParameters.IsLocalUserEnabled = this.enableLocalUser;
+            }
+            if(this.AllowedCopyScope != null)
+            {
+                createParameters.AllowedCopyScope = this.AllowedCopyScope;
             }
 
             var createAccountResponse = this.StorageClient.StorageAccounts.Create(