+++ src/EventHub/EventHub/ChangeLog.md
@@ -19,6 +19,10 @@
 -->
 ## Upcoming Release
 
+## Version 2.2.1
+* In the upcoming major breaking change release in October 2022, Az.EventHub would be migrating most cmdlets to a new format
+for a better powershell experience and as a result would witness breaking changes. Please refer our migration guide to know more https://go.microsoft.com/fwlink/?linkid=2204690.
+
 ## Version 2.2.0
 * Added -MinimumTlsVersion to New-AzEventHubNamespace and Set-AzEventHubNamespace
 * Added -SupportsScaling to New-AzEventHubCluster and Set-AzEventHubCluster to support self serve clusters